implement router and multipages with react router
