--- src/index.css
@@ -1,52 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-.calculatorContainer {
-  display: flex;
-  flex-direction: column;
-  margin: 200px;
-  width: 300px;
-  height: 600px;
-}
-
-.display {
-  display: flex;
-  background-color: grey;
-  color: antiquewhite;
-  justify-content: end;
-  height: 50px;
-  gap: 5px;
-}
-
-.buttonsContainer {
-  display: flex;
-  width: 100%;
-  flex-wrap: wrap;
-}
-
-.characterButton {
-  background-color: lightgrey;
-}
-
-.operationButton {
-  background-color: rgb(238, 129, 27);
-}
-
-.button {
-  text-align: center;
-  border-style: solid;
-  border-color: grey;
-  border-width: 1px;
-  height: 50px;
-}
-
-.s1 {
-  width: 25%;
-}
-
-.s2 {
-  width: 50%;
-}
