+++ src/pages/Quote.js
@@ -1,0 +1,10 @@
+const Quote = () => (
+  <div className="quotePageContainer">
+    <p className="quote">
+      “mathematics is not about numbers, equations, computations or
+      algorithms; its about understanding.” - William Thurston.
+    </p>
+  </div>
+);
+
+export default Quote;
