+++ src/styles/index.css
@@ -1,0 +1,86 @@
+@import url('https://fonts.googleapis.com/css2?family=Comic+Neue:wght@400;700&display=swap');
+
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+  font-family: 'Comic Neue', cursive;
+}
+
+header {
+  display: flex;
+  align-items: center;
+  margin: 20px 50px;
+  width: auto;
+  justify-content: space-between;
+}
+
+nav {
+  display: flex;
+  gap: 5px;
+}
+
+hr {
+  border: 1px solid black;
+  background-color: gray;
+}
+
+.display {
+  display: flex;
+  background-color: grey;
+  color: antiquewhite;
+  justify-content: flex-end;
+  height: 50px;
+  gap: 5px;
+}
+
+.buttonsContainer {
+  display: flex;
+  width: 100%;
+  flex-wrap: wrap;
+}
+
+.characterButton {
+  background-color: lightgrey;
+}
+
+.operationButton {
+  background-color: rgb(238, 129, 27);
+}
+
+.button {
+  text-align: center;
+  border-style: solid;
+  border-color: grey;
+  border-width: 1px;
+  height: 50px;
+}
+
+.s1 {
+  width: 25%;
+}
+
+.s2 {
+  width: 50%;
+}
+
+.homePageContainer {
+  display: flex;
+  flex-direction: column;
+  margin: 50px;
+  gap: 30px;
+}
+
+.quotePageContainer {
+  display: flex;
+  flex-direction: column;
+  margin: 100px auto;
+  justify-content: center;
+  align-items: center;
+  padding: 100px;
+}
+
+.quote {
+  font-size: xx-large;
+  text-align: center;
+}
